Finished the first version of the Dart implementation.
It has been implemented following the ideas of the Ruby version.

The Dart version is a little under 10x slower than the Ruby version.

The Dart version does some more work during the mask method because
of some lack of support for matching a RegExp after a certain position
on a string.